CHOOSE_CORRECT_ANSWER MC done
